Refactor formData formatting into helper function
