+++ src/pages/NewPost/NewPostPage.jsx
@@ -3,6 +3,7 @@ import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { createPost } from "../../api";
 import PostForm from "../../components/PostForm";
+import { createFormData } from "../../helpers";
 
 export default function NewPostPage() {
   const navigate = useNavigate();
@@ -12,16 +13,7 @@ export default function NewPostPage() {
   const handleCreatePost = async (values) => {
     setIsSubmitting(true);
     try {
-      const formData = new FormData();
-      formData.append("title", values.title);
-      formData.append("content", values.content);
-      formData.append("published", values.published || false);
-      if (values.coverImage) {
-        formData.append("coverImage", values.coverImage);
-      }
-      const tagsString = values.tags.join(",");
-      formData.append("tags", tagsString);
-
+      const formData = createFormData(values);
       await createPost(formData);
       navigate("/");
     } catch (error) {
@@ -32,7 +24,6 @@ export default function NewPostPage() {
       setIsSubmitting(false);
     }
   };
-
 
   return <PostForm onSubmit={handleCreatePost} isSubmitting={isSubmitting} />;
 }
